Block editor: separate styles store
