+++ packages/block-editor/src/store/reducer.js
@@ -2011,27 +2011,6 @@ export function openedBlockSettingsMenu( state = null, action ) {
 }
 
 /**
- * Reducer returning a map of style IDs to style overrides.
- *
- * @param {Map}    state  Current state.
- * @param {Object} action Dispatched action.
- *
- * @return {Map} Updated state.
- */
-export function styleOverrides( state = new Map(), action ) {
-	switch ( action.type ) {
-		case 'SET_STYLE_OVERRIDE':
-			return new Map( state ).set( action.id, action.style );
-		case 'DELETE_STYLE_OVERRIDE': {
-			const newState = new Map( state );
-			newState.delete( action.id );
-			return newState;
-		}
-	}
-	return state;
-}
-
-/**
  * Reducer returning a map of the registered inserter media categories.
  *
  * @param {Array}  state  Current state.
@@ -2092,7 +2071,6 @@ const combinedReducers = combineReducers( {
 	temporarilyEditingFocusModeRevert,
 	blockVisibility,
 	blockEditingModes,
-	styleOverrides,
 	removalPromptData,
 	blockRemovalRules,
 	openedBlockSettingsMenu,
