+++ packages/block-editor/src/store/styles/index.js
@@ -1,0 +1,29 @@
+/**
+ * WordPress dependencies
+ */
+import { createReduxStore, registerStore } from '@wordpress/data';
+
+/**
+ * Internal dependencies
+ */
+import reducer from './reducer';
+import * as selectors from './selectors';
+import * as actions from './actions';
+
+// To do: make store private somehow.
+const STORE_NAME = 'core/block-editor/styles';
+
+export const storeConfig = {
+	reducer,
+	selectors,
+	actions,
+};
+
+/**
+ * Store definition for the block editor namespace.
+ *
+ * @see https://github.com/WordPress/gutenberg/blob/HEAD/packages/data/README.md#createReduxStore
+ */
+export const store = createReduxStore( STORE_NAME, storeConfig );
+
+registerStore( STORE_NAME, storeConfig );
